refactor(follow): CursorPageResult 위치 이동
커서 기반 페이지 결과를 담은 공통 dto 이므로 common 패키지로 이동

#27
